Remove reducer and transfer to form-hook
